RuleFunctionBulder
