+++ src/main/java/org/delia/rule/RuleFunctionBulder.java
@@ -3,9 +3,8 @@ package org.delia.rule;
 import org.delia.compiler.astx.XNAFMultiExp;
 import org.delia.type.DType;
 
-public interface RuleFunctionFactory {
+public interface RuleFunctionBulder {
 
-	void addBuilder(RuleFunctionBulder builder);
 	DRule createRule(XNAFMultiExp rfe, int index, DType dtype);
 
 }
